MPS: enable pushing models/losses/data to configured device
Going via the config module (used in any of the exposed entry points) will
ensure a valid compute device is configured.

In the past the code was not respecting the set device everywhere.
I suspect running on CPU via config on a machine with a cuda device would
not have worked.

Note: On windows and linux (no mps) the behavior should be unchanged even
when using pytorch-3dunet as a library. This is ensure by the code paths
involving `legacy_default_device`. (So currently instantiating anything
on a mac directly would result in running on the cpu -- as before).